+++ src/main/java/application/poudlard/dao/EtudiantDAO.java
@@ -11,16 +11,16 @@ import java.util.List;
 import java.util.Optional;
 
 @Repository
-public interface EtudiantDAO  extends JpaRepository<Etudiant, Integer> {
-    boolean existsById(Integer id);
+public interface EtudiantDAO extends JpaRepository<Etudiant, Long> {
+    boolean existsById(Long id);
 
     @Query("SELECT e.notes FROM Etudiant e WHERE e.idEtudiant = :etudiantId")
-    List<Note> getNotes(@Param("etudiantId") Integer etudiantId);
+     List<Note> getNotes(@Param("etudiantId") Long etudiantId);
 
     boolean existsByEmail(String email);
     Optional<Etudiant> findByEmail(String email);
 
     @Query("SELECT e FROM Etudiant e JOIN e.cours c WHERE c.idCours = :idCours")
-    List<Etudiant> findEtudiantsByCours(@Param("idCours") int idCours);
+    List<Etudiant> findEtudiantsByCours(@Param("idCours") Long idCours);
 
 }
